refactor(codegen): extract dependencies into plugin folder
