+++ scripts/codegen/src/core/registry.ts
@@ -1,3 +1,4 @@
+import { dependenciesPlugin } from "../plugins/dependencies/index.ts";
 import { motionPlugin } from "../plugins/motion/index.ts";
 import { tokensPlugin } from "../plugins/tokens/index.ts";
 import { variantsPlugin } from "../plugins/variants/index.ts";
@@ -8,4 +9,9 @@ import type { SubstratePlugin } from "./plugin.ts";
  * contributions are visible to later plugins' analyze() AND to every
  * plugin's emit(). Plugins never import one another.
  */
-export const PLUGINS: readonly SubstratePlugin[] = [motionPlugin, tokensPlugin, variantsPlugin];
+export const PLUGINS: readonly SubstratePlugin[] = [
+  motionPlugin,
+  tokensPlugin,
+  variantsPlugin,
+  dependenciesPlugin,
+];
