+++ scripts/codegen/src/plugins/dependencies/index.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+import type { SubstratePlugin } from "../../core/plugin.ts";
+
+export const dependenciesPlugin: SubstratePlugin = {
+  name: "dependencies",
+  schema: {
+    atomic: { dependencies: z.array(z.string()).optional() },
+    composite: { dependencies: z.array(z.string()).optional() },
+  },
+  emit: {},
+};
