+++ scripts/codegen/src/semantic-checks.test.ts
@@ -7,7 +7,6 @@ import {
   checkConstraintsAgainstCoverage,
   checkConstraintsAgainstExamples,
   checkCoverageShape,
-  checkDependencyCycles,
   checkElementByProp,
   checkEvents,
   checkEventsRuntimeSupport,
@@ -236,35 +235,6 @@ describe("checkVocabulary", () => {
     });
     const issues = checkVocabulary(spec, vocabulary);
     expect(issues[0]?.message).toMatch(/typo of the canonical prop 'loading'/);
-  });
-});
-
-describe("checkDependencyCycles", () => {
-  test("flags a direct A → A cycle", () => {
-    const issues = checkDependencyCycles({ depsByName: new Map([["a", ["a"]]]) });
-    expect(issues).toHaveLength(1);
-  });
-
-  test("flags an indirect A → B → A cycle", () => {
-    const issues = checkDependencyCycles({
-      depsByName: new Map([
-        ["a", ["b"]],
-        ["b", ["a"]],
-      ]),
-    });
-    expect(issues.length).toBeGreaterThanOrEqual(1);
-  });
-
-  test("passes an acyclic graph", () => {
-    expect(
-      checkDependencyCycles({
-        depsByName: new Map([
-          ["a", ["b", "c"]],
-          ["b", ["c"]],
-          ["c", []],
-        ]),
-      }),
-    ).toEqual([]);
   });
 });
